@@ -4,15 +4,19 @@
   "private": true,
   "homepage": "https://GarenCrownguard.github.io/",
   "dependencies": {
+    "@chakra-ui/react": "^2.2.4",
+    "@emotion/react": "^11.10.0",
+    "@emotion/styled": "^11.10.0",
+    "framer-motion": "^4.1.17",
     "prop-types": "^15.7.2",
-    "react": "^16.13.1",
+    "react": "^18.2.0",
     "react-burger-menu": "^2.6.13",
     "react-collapsible-content": "^1.0.0",
-    "react-dom": "^16.8.6",
+    "react-dom": "^18.2.0",
     "react-jss": "^10.4.0",
     "react-redux": "^8.0.2",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
+    "react-scripts": "4.0.3",
     "react-svg-line-chart": "^2.0.1",
     "redux": "^4.2.0",
     "redux-devtools-extension": "^2.13.9",
